Refactor: Update Saldo Feature with New Store Opening Logic
- Modified `OpenStoreUseCase` to accept `Int` instead of `Double` for the initial amount.
- Updated `MainScreen` to:
    - Collect the `open` state from `SaldoViewModel`.
    - Show a "Buka Toko" button to trigger store opening.
    - Show a loading indicator during store opening.
    - Display a Toast message for errors.
    - Navigate to the next screen after successfully opening the store.
    - Modified shift Radio button selection to "P" and "S" for Pagi and Sore.
- Added `openStore` function to `SaldoViewModel` to handle the store opening process.
- Changed `SaldoViewModel` to only depends on CheckOpen and OpenStore UseCase.
- Added `Initial` state in `ResultState`.
- Removed unused dependency in `ViewModelModule` and `UseCaseModule`.
- Modified `SaldoPage` to:
    - Handle `Initial` state.
    - Call `checkSaldo` if the state is error.
- Modified `SaldoRequest` to accept `Int` instead of `Double` for the initial amount.
- Added `@SerializedName` for `hasOpenSaldo` key in `SaldoResponse`.
- Updated project configuration.

+++ app/src/main/java/cloud/hendra/danantara/domain/usecase/saldo/OpenStoreUseCase.kt
@@ -4,6 +4,6 @@ import cloud.hendra.danantara.data.repository.SaldoRepository
 import cloud.hendra.danantara.domain.model.SaldoRequest
 
 class OpenStoreUseCase(private val repository: SaldoRepository) {
-    suspend operator fun invoke(shift: String, awal: Double) =
+    suspend operator fun invoke(shift: String, awal: Int) =
         repository.openSaldo(SaldoRequest(shift, awal))
 }
